@@ -1,34 +1,266 @@
-# Copyright 2023–2025 Google LLC
-#
-# Licensed under the Apache License, Version 2.0 (the "License");
-# you may not use this file except in compliance with the License.
-# You may obtain a copy of the License at
-#
-#    https://www.apache.org/licenses/LICENSE-2.0
-#
-# Unless required by applicable law or agreed to in writing, software
-# distributed under the License is distributed on an "AS IS" BASIS,
-# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-# See the License for the specific language governing permissions and
-# limitations under the License.
-
 [build-system]
-requires = ["setuptools"]
-build-backend = "setuptools.build_meta"
+requires = ["hatchling"]
+build-backend = "hatchling.build"
+
+[tool.hatch.version]
+path = "src/MaxText/__init__.py"
 
 [project]
-name = "MaxText"
-version = "0.1.0"
-description = "MaxText is a high performance, highly scalable, open-source LLM written in pure Python/Jax and targeting Google Cloud TPUs and GPUs for training and inference"
+name = "maxtext"
+description = "MaxText is a simple, performant and scalable Jax LLM!"
+dynamic = ["version"]
+requires-python = ">=3.12"
 readme = "README.md"
-requires-python = ">=3.10"
-license = {file = "LICENSE"}
-dynamic = ["dependencies"]
+license = "Apache-2.0"
+keywords = ["llm", "jax", "llama", "mistral", "mixtral", "gemma", "deepseek"]
+classifiers = [
+    "Development Status :: 4 - Beta",
+    "Programming Language :: Python",
+    "Programming Language :: Python :: 3.12",
+    "Programming Language :: ML",
+    "Environment :: GPU"
+]
+dependencies = [
+    "absl-py>=2.3.1",
+    "aiofiles>=24.1.0",
+    "aiohappyeyeballs>=2.6.1",
+    "aiohttp>=3.12.15",
+    "aiosignal>=1.4.0",
+    "annotated-types>=0.7.0",
+    "antlr4-python3-runtime>=4.9.3",
+    "anyio>=4.10.0",
+    "aqtp>=0.9.0",
+    "array-record>=0.8.1",
+    "astroid>=3.3.11",
+    "astunparse>=1.6.3",
+    "attrs>=25.3.0",
+    "auditwheel>=6.4.1",
+    "black>=24.10.0",
+    "blobfile>=3.1.0",
+    "build>=1.2.2.post1",
+    "cachetools>=5.5.2",
+    "certifi>=2025.8.3",
+    "cffi>=1.17.1 ; platform_python_implementation == 'PyPy'",
+    "cfgv>=3.4.0",
+    "charset-normalizer>=3.4.3",
+    "chex>=0.1.90",
+    "click>=8.2.1",
+    "cloud-accelerator-diagnostics>=0.1.1",
+    "cloud-tpu-diagnostics>=0.1.5",
+    "cloudpickle>=3.1.1",
+    "clu>=0.0.12",
+    "colorama>=0.4.6",
+    "contourpy>=1.3.2",
+    "coverage>=7.10.6",
+    "cycler>=0.12.1",
+    "decorator>=5.2.1",
+    "dill>=0.3.8",
+    "distlib>=0.4.0",
+    "dm-tree>=0.1.9",
+    "docstring-parser>=0.17.0",
+    "editdistance>=0.8.1",
+    "einops>=0.8.1",
+    "etils>=1.13.0",
+    "evaluate>=0.4.5",
+    "execnet>=2.1.1",
+    "fastapi>=0.116.1",
+    "filelock>=3.18.0",
+    "flatbuffers>=25.2.10",
+    "flax>=0.11.2",
+    "fonttools>=4.59.0",
+    "frozenlist>=1.7.0",
+    "fsspec>=2025.7.0",
+    "gast>=0.6.0",
+    "gcsfs>=2025.7.0",
+    "google-api-core>=2.25.1",
+    "google-api-python-client>=2.181.0",
+    "google-auth-httplib2>=0.2.0",
+    "google-auth-oauthlib>=1.2.2",
+    "google-auth>=2.40.3",
+    "google-cloud-aiplatform>=1.111.0",
+    "google-cloud-appengine-logging>=1.6.2",
+    "google-cloud-audit-log>=0.3.2",
+    "google-cloud-bigquery>=3.36.0",
+    "google-cloud-core>=2.4.3",
+    "google-cloud-logging>=3.12.1",
+    "google-cloud-monitoring>=2.27.2",
+    "google-cloud-resource-manager>=1.14.2",
+    "google-cloud-storage>=2.19.0",
+    "google-crc32c>=1.7.1",
+    "google-genai>=1.33.0",
+    "google-pasta>=0.2.0",
+    "google-resumable-media>=2.7.2",
+    "google-tunix>=0.1.0",
+    "googleapis-common-protos>=1.70.0",
+    "grain>=0.2.12",
+    "grpc-google-iam-v1>=0.14.2",
+    "grpcio-status>=1.71.2",
+    "grpcio>=1.74.0",
+    "gviz-api>=1.10.0",
+    "h11>=0.16.0",
+    "h5py>=3.14.0",
+    "hf-xet>=1.1.9 ; platform_machine == 'aarch64' or platform_machine == 'amd64' or platform_machine == 'arm64' or platform_machine == 'x86_64'",
+    "httpcore>=1.0.9",
+    "httplib2>=0.30.0",
+    "httpx>=0.28.1",
+    "huggingface-hub>=0.34.4",
+    "humanize>=4.13.0",
+    "hypothesis>=6.136.4",
+    "identify>=2.6.14",
+    "idna>=3.10",
+    "immutabledict>=4.2.1",
+    "importlab>=0.8.1",
+    "importlib-metadata>=8.7.0",
+    "importlib-resources>=6.5.2",
+    "iniconfig>=2.1.0",
+    "isort>=6.0.1",
+    "jax>=0.7.0",
+    "jaxlib>=0.7.0",
+    "jaxtyping>=0.3.2",
+    "jinja2>=3.1.6",
+    "joblib>=1.5.2",
+    "jsonlines>=4.0.0",
+    "keras>=3.11.3",
+    "kiwisolver>=1.4.8",
+    "libclang>=18.1.1",
+    "libcst>=1.8.2",
+    "libtpu>=0.0.19 ; platform_machine == 'x86_64' and sys_platform == 'linux'",
+    "lxml>=6.0.1",
+    "markdown-it-py>=3.0.0",
+    "markdown>=3.9",
+    "markupsafe>=3.0.2",
+    "matplotlib>=3.10.3",
+    "mccabe>=0.7.0",
+    "mdurl>=0.1.2",
+    "ml-collections>=1.1.0",
+    "ml-dtypes>=0.5.3",
+    "ml-goodput-measurement>=0.0.14",
+    "more-itertools>=10.8.0",
+    "mpmath>=1.3.0",
+    "msgpack>=1.1.1",
+    "msgspec>=0.19.0",
+    "multidict>=6.6.4",
+    "multiprocess>=0.70.16",
+    "mypy-extensions>=1.1.0",
+    "namex>=0.1.0",
+    "nest-asyncio>=1.6.0",
+    "networkx>=3.5",
+    "ninja>=1.13.0",
+    "nltk>=3.9.1",
+    "nodeenv>=1.9.1",
+    "numpy>=2.0.2",
+    "oauthlib>=3.3.1",
+    "omegaconf>=2.3.0",
+    "opentelemetry-api>=1.36.0",
+    "opt-einsum>=3.4.0",
+    "optax>=0.2.5",
+    "optree>=0.17.0",
+    "orbax-checkpoint>=0.11.24",
+    "packaging>=25.0",
+    "pandas>=2.3.2",
+    "parameterized>=0.9.0",
+    "pathspec>=0.12.1",
+    "pathwaysutils>=0.1.2",
+    "pillow>=11.3.0",
+    "platformdirs>=4.4.0",
+    "pluggy>=1.6.0",
+    "portpicker>=1.6.0",
+    "pre-commit>=4.3.0",
+    "prometheus-client>=0.22.1",
+    "promise>=2.3",
+    "propcache>=0.3.2",
+    "proto-plus>=1.26.1",
+    "protobuf>=5.29.5",
+    "psutil>=7.0.0",
+    "pyarrow>=21.0.0",
+    "pyasn1-modules>=0.4.2",
+    "pyasn1>=0.6.1",
+    "pycnite>=2024.7.31",
+    "pycparser>=2.22 ; platform_python_implementation == 'PyPy'",
+    "pycryptodomex>=3.23.0",
+    "pydantic-core>=2.33.2",
+    "pydantic>=2.11.7",
+    "pydot>=4.0.1",
+    "pyelftools>=0.32",
+    "pyglove>=0.4.5",
+    "pygments>=2.19.2",
+    "pyink>=24.10.1",
+    "pylint>=3.3.8",
+    "pyparsing>=3.2.3",
+    "pyproject-hooks>=1.2.0",
+    "pytest-xdist>=3.8.0",
+    "pytest>=8.4.1",
+    "python-dateutil>=2.9.0.post0",
+    "pytype>=2024.10.11",
+    "pytz>=2025.2",
+    "pyyaml>=6.0.2",
+    "qwix>=0.1.0",
+    "regex>=2025.9.1",
+    "requests-oauthlib>=2.0.0",
+    "requests>=2.32.5",
+    "rich>=14.1.0",
+    "rsa>=4.9.1",
+    "safetensors>=0.6.2",
+    "scipy>=1.16.0",
+    "sentencepiece>=0.2.1",
+    "seqio>=0.0.20",
+    "setuptools>=80.9.0",
+    "shapely>=2.1.1",
+    "shortuuid>=1.0.13",
+    "simple-parsing>=0.1.7",
+    "simplejson>=3.20.1",
+    "six>=1.17.0",
+    "sniffio>=1.3.1",
+    "sortedcontainers>=2.4.0",
+    "starlette>=0.47.3",
+    "sympy>=1.14.0",
+    "tabulate>=0.9.0",
+    "tenacity>=9.1.2",
+    "tensorboard-data-server>=0.7.2",
+    "tensorboard-plugin-profile>=2.13.0",
+    "tensorboard>=2.19.0",
+    "tensorboardx>=2.6.4",
+    "tensorflow-datasets>=4.9.9",
+    "tensorflow-metadata>=1.17.2",
+    "tensorflow-text>=2.19.0",
+    "tensorflow>=2.19.1",
+    "tensorstore>=0.1.76",
+    "termcolor>=3.1.0",
+    "tiktoken>=0.11.0",
+    "tokenizers>=0.22.0",
+    "toml>=0.10.2",
+    "tomlkit>=0.13.3",
+    "toolz>=1.0.0",
+    "tqdm>=4.67.1",
+    "transformers>=4.56.1",
+    "treescope>=0.1.10",
+    "typing-extensions>=4.14.1",
+    "typing-inspection>=0.4.1",
+    "tzdata>=2025.2",
+    "uritemplate>=4.2.0",
+    "urllib3>=2.5.0",
+    "uvicorn>=0.35.0",
+    "virtualenv>=20.34.0",
+    "wadler-lindig>=0.1.7",
+    "websockets>=15.0.1",
+    "werkzeug>=3.1.3",
+    "wheel>=0.45.1",
+    "wrapt>=1.17.3",
+    "xxhash>=3.5.0",
+    "yarl>=1.20.1",
+    "zipp>=3.23.0",
+    "zstandard>=0.23.0",
+]
 
-[tool.setuptools.dynamic]
-dependencies = {file = ["requirements.txt"]}
+[project.urls]
+Repository = "https://github.com/AI-Hypercomputer/maxtext.git"
+"Bug Tracker" = "https://github.com/AI-Hypercomputer/maxtext/issues"
 
-[tool.setuptools.packages.find]  
-where = ["MaxText"]
-include = ["MaxText*"]
-exclude = ["MaxText.tests.*"]
+[tool.hatch.metadata]
+allow-direct-references = true
+
+[tool.hatch.build.targets.wheel]
+packages = ["src/MaxText", "src/install_maxtext_extra_deps"]
+
+[project.scripts]
+install_maxtext_github_deps = "install_maxtext_extra_deps.install_github_deps:main"
