@@ -12,12 +12,58 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
+"""This file contains common utilities for MaxText benchmarking.
+
+It includes:
+  - The `MaxTextModel` dataclass for defining model configurations.
+  - Helper functions like `str2bool` for parsing arguments.
+"""
+
 import dataclasses
-import math
 import typing
+
+
+def str2bool(v: str) -> bool:
+  """Convert a string of truth to True or False.
+
+  Args:
+    - v (str):
+      - True values are 'y', 'yes', 't', 'true', and '1';
+      - False values are 'n', 'no', 'f', 'false', and '0'.
+
+  Returns:
+    bool: True or False
+
+  Raises:
+    ValueError if v is anything else.
+  """
+  v = v.lower()
+  true_values = ["y", "yes", "t", "true", "1"]
+  false_values = ["n", "no", "f", "false", "0"]
+  if v in true_values:
+    return True
+  elif v in false_values:
+    return False
+  else:
+    raise ValueError(f"Invalid value '{v}'!")
+
 
 @dataclasses.dataclass
 class MaxTextModel:
+  """A dataclass for representing a MaxText model configuration for benchmarking.
+
+  Attributes:
+    model_name: The user-facing name of the model configuration.
+    model_type: The specific model variant to be run (e.g., '7b', '13b').
+    tuning_params: A dictionary of hyperparameters and settings to override
+      the base configuration.
+    xla_flags: A string of XLA flags to be used for the model compilation.
+    pathways_tuning_params: An optional dictionary of tuning parameters specific
+      to Pathways execution.
+    pathways_xla_flag_options: An optional dictionary to customize XLA flags
+      for Pathways, allowing for additions or removals.
+  """
+
   model_name: str
   model_type: str
   tuning_params: dict[str, typing.Any]
