@@ -12,25 +12,29 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
+"""A recipe for running an elastic training benchmark with disruptions.
+
+This script configures and launches a MaxText workload on a GKE cluster using XPK,
+and then introduces disruptions (e.g., killing pods) to test the resilience
+and recovery capabilities of the training job. It can be configured to run
+with both Pathways and McJAX to compare their elastic training behavior.
+"""
+
 import os
-import sys
 
 import args_helper as helper
 
-parent_dir = os.path.abspath(os.path.join(os.path.dirname(__file__), ".."))
-sys.path.append(parent_dir)
-
-from disruption_management.disruption_handler import DisruptionConfig
-from disruption_management.disruption_handler import DisruptionMethod
-from disruption_management.disruption_handler import MCJAX_STANDARD_TARGET_POD_REGEX_SUFFIX
-from disruption_management.disruption_handler import MCJAX_WORKER_CONTAINER_NAME
-from disruption_management.disruption_handler import PATHWAYS_STANDARD_TARGET_POD_REGEX_SUFFIX
-from disruption_management.disruption_handler import PATHWAYS_WORKER_CONTAINER_NAME
-from disruption_management.disruption_handler import TriggerType
-from maxtext_trillium_model_configs import MaxTextModel
-import maxtext_v5e_model_configs as v5e_model_configs
-import maxtext_xpk_runner as mxr
-from xpk_configs import XpkClusterConfig
+from benchmarks.disruption_management.disruption_handler import DisruptionConfig
+from benchmarks.disruption_management.disruption_handler import DisruptionMethod
+from benchmarks.disruption_management.disruption_handler import MCJAX_STANDARD_TARGET_POD_REGEX_SUFFIX
+from benchmarks.disruption_management.disruption_handler import MCJAX_WORKER_CONTAINER_NAME
+from benchmarks.disruption_management.disruption_handler import PATHWAYS_STANDARD_TARGET_POD_REGEX_SUFFIX
+from benchmarks.disruption_management.disruption_handler import PATHWAYS_WORKER_CONTAINER_NAME
+from benchmarks.disruption_management.disruption_handler import TriggerType
+from benchmarks.maxtext_trillium_model_configs import MaxTextModel
+from benchmarks import maxtext_v5e_model_configs as v5e_model_configs
+from benchmarks import maxtext_xpk_runner as mxr
+from benchmarks.xpk_configs import XpkClusterConfig
 
 PROXY_IMAGE = "us-docker.pkg.dev/cloud-tpu-v2-images/pathways/proxy_server"
 SERVER_IMAGE = "us-docker.pkg.dev/cloud-tpu-v2-images/pathways/server"
@@ -46,9 +50,7 @@ DEVICE_TYPE = "v6e-256"
 # Other parameters (MUST BE SET BY USER)
 XPK_PATH = "../xpk"  # We're running this script from the maxtext directory
 USER = os.environ["USER"]
-BASE_OUTPUT_DIRECTORY = (
-    f"gs://{USER}-{PROJECT}-{COUNTRY}/disruption_management/"
-)
+BASE_OUTPUT_DIRECTORY = f"gs://{USER}-{PROJECT}-{COUNTRY}/disruption_management/"
 MAX_RESTARTS = 10
 NUM_SLICES = 2
 BENCHMARK_STEPS = 101
@@ -85,7 +87,7 @@ def construct_disruption_configs(
           disruption_method=DisruptionMethod.SIGILL,
           target_pod_regex=target_pod_regex,
           worker_container_name=worker_container_name,
-      )
+      ),
   ]
 
 
@@ -107,7 +109,7 @@ def construct_workload_config_with_disruptions(
       pathways_config=pathways_config,
       xpk_path=XPK_PATH,
       num_steps=BENCHMARK_STEPS,
-      disruption_configs=construct_disruption_configs(pathways_config)
+      disruption_configs=construct_disruption_configs(pathways_config),
   )
 
 
@@ -123,12 +125,10 @@ def main() -> None:
   )
 
   # Handle command line arguments using args_helper
-  should_continue = helper.handle_cmd_args(
-      cluster_config, helper.DELETE, xpk_path=XPK_PATH
-  )
+  should_continue = helper.handle_cmd_args(cluster_config, helper.DELETE, xpk_path=XPK_PATH)
 
   if not should_continue:
-    return 0
+    return
 
   # Model Configuration - Using a simple default model for testing
   model = v5e_model_configs.llama3_1_8b_8192
@@ -137,7 +137,6 @@ def main() -> None:
       server_image=SERVER_IMAGE,
       proxy_server_image=PROXY_IMAGE,
       runner_image=RUNNER,
-
       # User can add additional flags here.
       server_flags="--enable_metrics_collection=false",
       proxy_flags="--enable_metrics_collection=false",
@@ -146,16 +145,12 @@ def main() -> None:
 
   # Pathways Workload Configuration with Disruption
   workload_configs = []
-  pathways_workload_config = construct_workload_config_with_disruptions(
-      cluster_config, model, pathways_config
-  )
+  pathways_workload_config = construct_workload_config_with_disruptions(cluster_config, model, pathways_config)
   workload_configs.append(pathways_workload_config)
 
   if COMPARE_WITH_MCJAX:
     # Add a workload config for MCJAX
-    mcjax_workload_config = construct_workload_config_with_disruptions(
-        cluster_config, model, None
-    )
+    mcjax_workload_config = construct_workload_config_with_disruptions(cluster_config, model, None)
     workload_configs.append(mcjax_workload_config)
 
   # Run the benchmark and use the returned disruption manager.
@@ -167,9 +162,7 @@ def main() -> None:
   # Wait for disruptions to complete
   disruption_manager.start_disruptions_and_wait_for_completion()
 
-  print(
-      "Elastic Training disruptions completed. Please check logs for results."
-  )
+  print("Elastic Training disruptions completed. Please check logs for results.")
 
 
 if __name__ == "__main__":
