@@ -13,140 +13,34 @@
 # limitations under the License.
 
 """Used to perf benchmarks between Pathways and McJax."""
-
-import datetime
-import sys
 import os
-import args_helper as helper
+import sys
 
 parent_dir = os.path.abspath(os.path.join(os.path.dirname(__file__), ".."))
 sys.path.append(parent_dir)
-
-import maxtext_trillium_model_configs as model_configs
-import maxtext_xpk_runner as mxr
-from xpk_configs import XpkClusterConfig
-
-
-PROXY_IMAGE = "us-docker.pkg.dev/cloud-tpu-v2-images/pathways/proxy_server"
-SERVER_IMAGE = "us-docker.pkg.dev/cloud-tpu-v2-images/pathways/server"
-RUNNER = "us-docker.pkg.dev/path/to/maxtext_runner"
-
-# Cluster Params
-CLUSTER = "v6e-256-cluster"
-PROJECT = "tpu-prod-env-cluster"
-ZONE = "us-east5-b"
-REGION = "us-east5"
-COUNTRY = "us"
-DEVICE_TYPE = "v6e-256"
-
-# Other parameters (MUST BE SET BY USER)
-XPK_PATH = os.path.join("~", "xpk")  # We're running this script from the maxtext directory
-USER = os.environ["USER"]
-BASE_OUTPUT_DIRECTORY = (
-    f"gs://{USER}-{PROJECT}-{COUNTRY}/pw_mcjax_benchmarking/"
-)
-
-BENCHMARK_STEPS = 20
+from . import args_helper as helper
+from . import user_configs
+from .runner_utils import generate_and_run_workloads
 
 
 def main() -> int:
-  # V6e cluster config
-  cluster_config = XpkClusterConfig(
-      cluster_name=CLUSTER,
-      project=PROJECT,
-      zone=ZONE,
-      device_type=DEVICE_TYPE,
-  )
-
-  # Handle command line arguments using args_helper
+  """Main program entry point"""
+  user_configs.USER_CONFIG.headless = False
   should_continue = helper.handle_cmd_args(
-      cluster_config, helper.DELETE, xpk_path=XPK_PATH
+      user_configs.USER_CONFIG.cluster_config, helper.DELETE, xpk_path=user_configs.USER_CONFIG.xpk_path
   )
 
   if not should_continue:
     return 0
 
-  models = {
-      "mcjax": [
-          # model_configs.llama3_1_8b_8192,
-          # model_configs.llama3_1_70b_8192,
-          # model_configs.llama3_1_405b_8192_fsdp_dcn,
-          # model_configs.llama2_70b_4096_real_data_long_run,
-      ],
-      "pathways": [
-          model_configs.llama3_1_8b_8192,
-          # model_configs.llama3_1_70b_8192,
-          # model_configs.llama3_1_405b_8192_fsdp_dcn,
-          # model_configs.llama2_70b_4096_real_data_long_run,
-      ]
-  }
-  pathways_config = mxr.PathwaysConfig(
-      server_image=SERVER_IMAGE,
-      proxy_server_image=PROXY_IMAGE,
-      runner_image=RUNNER,
-
-      # User can add additional flags here.
-      server_flags="",
-      proxy_flags="",
-      worker_flags="",
+  return_code = generate_and_run_workloads(
+      user_configs.USER_CONFIG,
+      user_configs.USER_CONFIG.num_slices_list,
+      user_configs.USER_CONFIG.benchmark_steps,
+      user_configs.USER_CONFIG.priority,
   )
-  num_slices_list = [
-      2
-  ]
 
-  xpk_workload_cmds = []
-  xpk_workload_names = []
-
-  for infra, model_list in models.items():
-    for model in model_list:
-      # Run workloads on the below clusters
-      for cluster_config in [
-          cluster_config,
-      ]:
-
-        # Make modifications to the model config here to add in any additional
-        # flags or changes to the model config.
-        model.tuning_params["use_vertex_tensorboard"] = True
-        model.tuning_params["vertex_tensorboard_project"] = PROJECT
-        model.tuning_params["vertex_tensorboard_region"] = REGION
-
-        # Run workloads in the following slice configurations
-        for num_slices in num_slices_list:
-          wl_config = mxr.WorkloadConfig(
-              model=model,
-              num_slices=num_slices,
-              device_type=cluster_config.device_type,
-              base_output_directory=BASE_OUTPUT_DIRECTORY
-              + f"{infra}_{num_slices}_slice_{DEVICE_TYPE}_{model.model_name}/",
-              max_restarts=0,
-              libtpu_type=None,
-              libtpu_nightly_version="",
-              base_docker_image=RUNNER if infra == "mcjax" else None,
-              pathways_config=pathways_config if infra == "pathways" else None,
-              xpk_path=XPK_PATH,
-              num_steps=BENCHMARK_STEPS,
-              priority="low",
-          )
-          command, name = mxr.generate_xpk_workload_cmd(
-              cluster_config=cluster_config, wl_config=wl_config
-          )
-
-          print(f"Name of the workload is: {name} \n")
-          xpk_workload_names.append(name)
-
-          print(f"XPK command to be used is: {command} \n")
-          xpk_workload_cmds.append(command)
-
-  for xpk_workload_name, xpk_workload_cmd in zip(
-      xpk_workload_names, xpk_workload_cmds
-  ):
-    timestamp = datetime.datetime.now().strftime("%Y-%m-%d %H:%M:%S")
-    print(f"[{timestamp}] Running workload: {xpk_workload_name} with command: {xpk_workload_cmd}")
-    return_code = mxr.run_command_with_updates(
-        xpk_workload_cmd, xpk_workload_name
-    )
-    if return_code != 0:
-      print(f"Unable to run xpk workload: {xpk_workload_name}")
+  return return_code
 
 
 if __name__ == "__main__":
